combine the api files
