--- api/v1/auth_endpoints.py
@@ -1,50 +1,0 @@
-from fastapi import APIRouter, HTTPException, Depends
-from fastapi.security import OAuth2PasswordRequestForm
-from datetime import timedelta
-# from api.auth import authenticate_user, create_access_token
-from src.auth import authenticate_user, create_access_token
-from pydantic import BaseModel
-
-router = APIRouter()
-
-# Token response schema
-class TokenResponse(BaseModel):
-    access_token: str
-    token_type: str
-
-# Keep your custom login request for other clients
-class LoginRequest(BaseModel):
-    username: str
-    password: str
-
-@router.post("/login", response_model=TokenResponse)
-async def login(form_data: OAuth2PasswordRequestForm = Depends()):
-    """
-    OAuth2 compatible token endpoint for Swagger UI.
-    Accepts form data with username and password.
-    """
-    user = authenticate_user(form_data.username, form_data.password)
-    if not user:
-        raise HTTPException(status_code=401, detail="Invalid username or password")
-
-    access_token_expires = timedelta(minutes=30)
-    access_token = create_access_token(
-        data={"sub": user["username"]}, expires_delta=access_token_expires
-    )
-    return {"access_token": access_token, "token_type": "bearer"}
-
-# Alternative endpoint for JSON-based login (optional)
-@router.post("/login-json", response_model=TokenResponse)
-async def login_json(form_data: LoginRequest):
-    """
-    JSON-based login endpoint for custom clients.
-    """
-    user = authenticate_user(form_data.username, form_data.password)
-    if not user:
-        raise HTTPException(status_code=401, detail="Invalid username or password")
-
-    access_token_expires = timedelta(minutes=30)
-    access_token = create_access_token(
-        data={"sub": user["username"]}, expires_delta=access_token_expires
-    )
-    return {"access_token": access_token, "token_type": "bearer"}
