+++ src/preprocessing.py
@@ -11,19 +11,14 @@ class DataPreprocessor:
     def preprocess_data(self, df):
         """Preprocess the data for survival analysis."""
 
-        # Drop duplicates
         df = df.drop_duplicates()
 
-        # Convert datetime columns
         df = self._process_datetime_features(df)
 
-        # Create derived features
         df = self._create_derived_features(df)
 
-        # Create survival target
         y = self._create_survival_target(df)
 
-        # Prepare features
         X = self._prepare_features(df)
         print("  ✅ Data Preprocessing completed.")
 
@@ -36,7 +31,6 @@ class DataPreprocessor:
         df['Order_Creation_DateTime'] = pd.to_datetime(df['Order_Creation_DateTime'])
         df['Acknowledgement_DateTime'] = pd.to_datetime(df['Acknowledgement_DateTime'])
 
-        # Extract date components
         df['Order_Creation_Day'] = df['Order_Creation_DateTime'].dt.day
         df['Order_Creation_Month'] = df['Order_Creation_DateTime'].dt.month
         df['Order_Creation_Year'] = df['Order_Creation_DateTime'].dt.year
@@ -52,15 +46,12 @@ class DataPreprocessor:
         """Create derived features."""
         print("  ➡️  Creating derived features...")
 
-        # Time to acknowledge
         df['Time_to_Acknowledge'] = (
             df['Acknowledgement_DateTime'] - df['Order_Creation_DateTime']
         ).dt.days
 
-        # Low lead time flag
         df['is_low_lead_time'] = (df['Lead_Time'] <= 4.5).astype(int)
 
-        # Censoring time
         df['Censoring_Time'] = df['Lead_Time'].fillna(
             (CUTOFF_DATE - df['Order_Creation_DateTime']).dt.days
         )
@@ -73,7 +64,6 @@ class DataPreprocessor:
         event_mask = df['Lead_Time'].notna()
         observed_time = df['Censoring_Time']
 
-        # ✅ FIX: use Surv.from_arrays instead of np.array
         y = Surv.from_arrays(event=event_mask.astype(bool),
                              time=observed_time.astype(float))
 
@@ -85,10 +75,8 @@ class DataPreprocessor:
         X = df[FEATURES]
         X_encoded = pd.get_dummies(X, drop_first=True)
 
-        # Ensure the index of X_encoded is the same as the original DataFrame
         X_encoded = X_encoded.set_index(df.index)
 
-        # Store training columns for consistency in prediction
         self.training_columns = X_encoded.columns.tolist()
 
         print(f"  ✅ Features encoded.")
