+++ src/schemas.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+class TokenResponse(BaseModel):
+    access_token: str
+    token_type: str
+
+class LoginRequest(BaseModel):
+    username: str
+    password: str
